# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for StreamShort, a short-form video streaming platform.",
        "title": "StreamShort API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/analytics/watch": {
            "post": {
                "description": "Record a watch event for an episode.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Record Watch",
                "parameters": [
                    {
                        "description": "Watch Event",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.WatchEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.WatchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.WatchEventRequest": {
            "type": "object",
            "properties": {
                "episode_id": {
                    "type": "string"
                },
                "watched_seconds": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.WatchResponse": {
            "type": "object",
            "properties": {
                "first_watch": {
                    "type": "boolean"
                },
                "success": {
                    "type": "boolean"
                },
                "tracked": {
                    "type": "boolean"
                }
            }
        },
...
    }
}